v8.0: package refactor
