+++ src/main/java/ru/yandex/practicum/filmorate/controller/ErrorHandler.java
@@ -1,5 +1,6 @@
-package ru.yandex.practicum.filmorate.exception;
+package ru.yandex.practicum.filmorate.controller;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -7,30 +8,35 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.server.ResponseStatusException;
+import ru.yandex.practicum.filmorate.exception.ErrorDto;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 
 import java.util.Date;
 
 @RestControllerAdvice("ru.yandex.practicum.filmorate.controller")
+@Slf4j
 public class ErrorHandler {
     @ExceptionHandler(ResponseStatusException.class)
     public ResponseEntity<String> handleException(ResponseStatusException e) {
+        log.error(e.getMessage(), e);
         return ResponseEntity
                 .status(e.getStatus())
                 .body(e.getMessage() + " " + e.getResponseHeaders());
     }
 
     @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<ErrorDTO> handleException(NotFoundException e) {
-        ErrorDTO errorDTO = new ErrorDTO();
+    public ResponseEntity<ErrorDto> handleException(NotFoundException e) {
+        ErrorDto errorDTO = new ErrorDto();
         errorDTO.setMessage(e.getMessage());
         errorDTO.setStatus(String.valueOf(e.getStatus().value()));
         errorDTO.setTime(new Date().toString());
-
+        log.error(e.getMessage(), e);
         return new ResponseEntity<>(errorDTO, e.getStatus());
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<String> handleException(MethodArgumentNotValidException e) {
+        log.error(e.getMessage(), e);
         return ResponseEntity
                 .status(HttpStatus.BAD_REQUEST)
                 .body(HttpStatus.BAD_REQUEST + " " + e.getFieldError().getDefaultMessage());
@@ -38,6 +44,7 @@ public class ErrorHandler {
 
     @ExceptionHandler(MethodArgumentTypeMismatchException.class)
     public ResponseEntity<String> handleException(MethodArgumentTypeMismatchException e) {
+        log.error(e.getMessage(), e);
         return ResponseEntity
                 .status(HttpStatus.BAD_REQUEST)
                 .body(HttpStatus.BAD_REQUEST + " bad string parameters " + e.getName() + "=" + e.getValue());
